# GGRC-5881 Fix performance for comment adding