--- a/src/ggrc/models/background_task.py
+++ b/src/ggrc/models/background_task.py
@@ -3,7 +3,9 @@
 
 """Module for ggrc background tasks."""
 
+import json
 import traceback
+import uuid
 from logging import getLogger
 from functools import wraps
 from time import time
@@ -27,6 +29,17 @@ from ggrc.models import reflection
 
 
 logger = getLogger(__name__)
+
+BANNED_HEADERS = {
+    "X-Appengine-Country",
+    "X-Appengine-Queuename",
+    "X-Appengine-Current-Namespace",
+    "X-Appengine-Taskname",
+    "X-Appengine-Tasketa",
+    "X-Appengine-Taskexecutioncount",
+    "X-Appengine-Taskretrycount",
+    "X-Task-Id",
+}
 
 
 class BackgroundTask(roleable.Roleable, base.ContextRBAC, Base, Stateful,
@@ -104,6 +117,11 @@ def _add_task_acl(task):
     clear_users_permission_cache([get_current_user().id])
 
 
+def collect_task_headers():
+  """Get headers required for appengine background task run."""
+  return Headers({k: v for k, v in request.headers if k not in BANNED_HEADERS})
+
+
 def create_task(name, url, queued_callback=None, parameters=None, method=None):
   """Create a enqueue a bacground task."""
   if not method:
@@ -118,21 +136,10 @@ def create_task(name, url, queued_callback=None, parameters=None, method=None):
   task.modified_by = get_current_user()
   _add_task_acl(task)
 
-  banned = {
-      "X-Appengine-Country",
-      "X-Appengine-Queuename",
-      "X-Appengine-Current-Namespace",
-      "X-Appengine-Taskname",
-      "X-Appengine-Tasketa",
-      "X-Appengine-Taskexecutioncount",
-      "X-Appengine-Taskretrycount",
-      "X-Task-Id",
-  }
-
   # schedule a task queue
   if getattr(settings, 'APP_ENGINE', False):
     from google.appengine.api import taskqueue
-    headers = Headers({k: v for k, v in request.headers if k not in banned})
+    headers = collect_task_headers()
     headers.add('X-Task-Id', task.id)
     taskqueue.add(
         queue_name="ggrc",
@@ -145,6 +152,37 @@ def create_task(name, url, queued_callback=None, parameters=None, method=None):
   elif queued_callback:
     queued_callback(task)
   return task
+
+
+def create_lightweight_task(name, url, queued_callback=None, parameters=None,
+                            method=None):
+  """Create background task.
+
+  This function create an app engine background task without handling
+  related BackgroundTask object.
+  """
+  if not method:
+    method = request.method
+
+  if not parameters:
+    parameters = {}
+
+  if getattr(settings, 'APP_ENGINE', False):
+    from google.appengine.api import taskqueue
+    taskqueue.add(
+        queue_name="ggrc",
+        url=url,
+        name="{}_{}".format(name + str(int(time())), uuid.uuid4()),
+        payload=json.dumps(parameters),
+        method=method,
+        headers=collect_task_headers()
+    )
+  elif queued_callback:
+    queued_callback(**parameters)
+  else:
+    raise ValueError(
+        "Either queued_callback should be provided or APP_ENGINE set to true."
+    )
 
 
 def make_task_response(id_):