--- a/src/ggrc/utils/__init__.py
+++ b/src/ggrc/utils/__init__.py
@@ -14,6 +14,8 @@ import json
 import re
 import sys
 
+from werkzeug import exceptions
+
 import sqlalchemy
 from sqlalchemy.orm import class_mapper
 
@@ -351,8 +353,14 @@ def validate_mimetype(accepted_mimetype):
 def get_task_attr(attr_name, params=None):
   """Get attribute value from params dict or request data."""
   attr = None
+
+  try:
+    json_data = flask.request.get_json(force=True)
+  except exceptions.BadRequest:
+    json_data = None
+
   if params and attr_name in params:
     attr = params.get(attr_name)
-  elif flask.request.json and attr_name in flask.request.json:
-    attr = flask.request.json.get(attr_name)
+  elif json_data and attr_name in json_data:
+    attr = json_data.get(attr_name)
   return attr