# Issue description

Time required for adding comments to assessment was increased in 1.19-Strawberry comparing with 1.18-Strawberry (from 5s to 7s).

# Steps to test the changes

1. Log in to GGRC-PERF https://ggrc-perf.appspot.com/audits/4567#!assessment
2. Open Assessment Info panel (title = 'Check timing') https://ggrc-perf.appspot.com/assessments/24643
3. Type comment to comment box and click Add button: saving time is 7.1 sec / 6.8 sec / 7.27 sec / 6.5 sec

*4. Select 'no' option from dropdown lca, type comment and click Save button: saving time is *7.12 sec / 7.15 sec / 7.17 sec

Actual Result: Performance for saving comment decreases and equals 7.1 sec / 6.8 sec / 7.27 sec / 6.5 sec

Expected Result: comment saving time should be ~ 5 sec

Load time: https://docs.google.com/spreadsheets/d/1gNrun5Gm-IOF5xXDB3YbRfRS_ru_ppctF2KiwDShDT0/edit#gid=2141217279

# Solution description

Performance degradation was caused by https://github.com/google/ggrc-core/pull/8091. It increase time of all background tasks processing.
When new comment is added, 2 requests are sent: POST on comments endpoint and PUT on assessments endpoint. It lead to run bg tasks for recalculating computed attributes, but time of this task starting was increased from 0.8s to 1.1s for PUT and from 0.3s to 0.6s on POST (measured on 24643 assessment).
In this PR bg task for calculation of computed attributes was replaced with "lightweight" version. Such task version is not handle ggrc BackgroundTask objects, it just add task into appengine taskqueue (BackgroundTask is not used for computed attributes).

This PR decrease time of comments adding to 5s.

# Sanity checklist

- [ ] I have clicked through the app to make sure my changes work and not break the app.
- [x] I have applied the correct milestone and labels.
- [x] My changes fix the issue described in the description (and do nothing else). 🤞
- [ ] My changes are covered by tests.
- [x] My changes follow our [performance guidelines](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/performance.rst).
- [x] My changes follow our [js](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/javascript.rst) and/or [python](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/python.rst) guidelines.
- [x] My commits follow our [commit guidelines](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/git/how_to_write_a_commit_message.rst).


# PR Review checklist

- [x] The changes fix the issue and don't cause any apparent regressions.
- [x] Labels and milestone are correctly set.
- [x] The solution description matches the changes in the code.
- [x] There is no apparent way to improve the performance & design of the new code.
- [x] The pull request is opened against the correct base branch.
- [ ] Upon merging, the Jira ticket's fixversion is correctly set and the ticket is moved to "QA - In Progress".