@mashabarashko 2018-08-21
@IlyaTsupryk Most part of changes look good from my side. Timings checked locally on ggrc-test db dump. Timings before changes:
<img width="678" alt="screen shot 2018-08-21 at 3 56 19 pm" src="https://user-images.githubusercontent.com/2690200/44405552-32a74080-a562-11e8-8f7c-543375fc557e.png">
Timings after changes:
<img width="694" alt="screen shot 2018-08-21 at 3 56 28 pm" src="https://user-images.githubusercontent.com/2690200/44405558-3935b800-a562-11e8-8866-5d1f5cdc4410.png">

Got an error when try to call `/admin/compute_attributes`:
<img width="733" alt="screen shot 2018-08-21 at 4 50 56 pm" src="https://user-images.githubusercontent.com/2690200/44405654-78640900-a562-11e8-9938-bc8ae8df328c.png">
Could you please investigate?