@IlyaTsupryk 2018-08-21
@mashabarashko Thank you for a note. It was fixed.